Always fill out PR title after conflict resolution
Before this change, cherry-picker used to loose the backported commit
message and didn't pass anything to the `push_to_remote()` method which
resulted in sending an API request to GitHub that had an empty title
upon resume. This caused GitHub respond with an error.

This code path was only problematic for people having a GitHub token in
their env and therefore enjoying the benefits of the PR autocreation.

After this change, cherry-picker picks up the commit message correctly
and extracts the proper PR title and body out of it.